+++ tinlake-ui/components/OnboardModal/index.tsx
@@ -8,7 +8,7 @@ import { PoolData as PoolDataV3, PoolState } from '../../ducks/pool'
 import { PoolsState } from '../../ducks/pools'
 import { PoolLink } from '../PoolLink'
 import { FormModal, InvestmentSteps } from './styles'
-import { AddressStatus } from '@centrifuge/onboard-api/src/controllers/address.controller'
+import type { AddressStatus } from '@centrifuge/onboard-api/src/controllers/types'
 
 interface Props {
   anchor?: React.ReactNode
